--- src/Model/AbstractTranslatable.php
@@ -1,47 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/*
- * This file is part of the Sonata Project package.
- *
- * (c) Thomas Rabaix <thomas.rabaix@sonata-project.org>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Sonata\TranslationBundle\Model;
-
-/**
- * NEXT_MAJOR: Remove this file.
- *
- * If you don't want to extend this class, you can use Translatable trait instead.
- *
- * @author Nicolas Bastien <nbastien.pro@gmail.com>
- *
- * @deprecated since version 2.x, to be removed in 3.0.
- */
-abstract class AbstractTranslatable
-{
-    /**
-     * @var string|null
-     */
-    protected $locale;
-
-    /**
-     * @param string $locale
-     */
-    public function setLocale($locale): void
-    {
-        $this->locale = $locale;
-    }
-
-    /**
-     * @return string|null
-     */
-    public function getLocale()
-    {
-        return $this->locale;
-    }
-}
