--- tests/Fixtures/Traits/ModelPersonalTranslatable.php
@@ -1,34 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/*
- * This file is part of the Sonata Project package.
- *
- * (c) Thomas Rabaix <thomas.rabaix@sonata-project.org>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Sonata\TranslationBundle\Tests\Fixtures\Traits;
-
-use Doctrine\Common\Collections\ArrayCollection;
-use Sonata\TranslationBundle\Model\Gedmo\AbstractPersonalTranslation;
-use Sonata\TranslationBundle\Model\Gedmo\TranslatableInterface;
-use Sonata\TranslationBundle\Traits\Gedmo\PersonalTranslatableTrait;
-
-class ModelPersonalTranslatable implements TranslatableInterface
-{
-    use PersonalTranslatableTrait;
-
-    /**
-     * @var ArrayCollection<array-key, AbstractPersonalTranslation>
-     */
-    protected $translations;
-
-    public function __construct()
-    {
-        $this->translations = new ArrayCollection();
-    }
-}
